+++ packages/api/src/controllers/TransactionsController.js
@@ -4,7 +4,6 @@ const TransactionsDAO = require("../dao/TransactionsDAO");
 
 class TransactionsController {
     constructor(client, knex) {
-        this.knex = knex
         this.client = client
         this.transactionsDAO = new TransactionsDAO(knex)
     }
@@ -12,21 +11,7 @@ class TransactionsController {
     getTransactions = async (request, response) => {
         const {from, to} = request.query
 
-        const rows = await this.knex
-            .select('state_transitions.hash as hash', 'state_transitions.data as data', 'state_transitions.type as type',
-                'state_transitions.index as index', 'blocks.height as block_height', 'blocks.timestamp as timestamp')
-            .from('state_transitions')
-            .leftJoin('blocks', 'blocks.hash', 'state_transitions.block_hash')
-            .where((builder) => {
-                if (from && to) {
-                    builder.where('block_height', '<', to);
-                    builder.where('block_height', '>', from);
-                }
-            })
-            .limit(30)
-            .orderBy('blocks.height', 'desc')
-
-        const transactions = rows.map((row) => Transaction.fromRow(row))
+        const transactions = await this.transactionsDAO.getTransactions(from, to)
 
         response.send(transactions);
     }
